ci: enforce MSRV and warning-free Clippy

+++ apps/desktop/src-tauri/src/diag.rs
@@ -116,7 +116,7 @@ fn iso8601_now() -> String {
     // public-domain days_from_civil inverse. Correct across leap years.
     let z = days as i64 + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
